chore(web): list type refactor
